refactor(racing-game): 패키지 분리
도메인 관련 기능은 domain 패키지로 UI 관련한 객체는 view 패키지로 이동
