+++ src/main/kotlin/week1/step4/racinggame/domain/RandomNumber.kt
@@ -1,4 +1,4 @@
-package week1.step4.racinggame
+package week1.step4.racinggame.domain
 
 @JvmInline
 value class RandomNumber(val value: Int) {
