+++ src/main/kotlin/week1/step4/racinggame/view/InputView.kt
@@ -1,4 +1,6 @@
-package week1.step4.racinggame
+package week1.step4.racinggame.view
+
+import week1.step4.racinggame.domain.CarName
 
 class InputView {
     companion object {
